Move rpc code out of state machine processor
