+++ mqtt-meta/src/main/java/org/apache/rocketmq/mqtt/meta/raft/processor/CounterStateProcessor.java
@@ -24,6 +24,7 @@ import com.google.protobuf.ByteString;
 import org.apache.rocketmq.mqtt.common.model.consistency.ReadRequest;
 import org.apache.rocketmq.mqtt.common.model.consistency.Response;
 import org.apache.rocketmq.mqtt.common.model.consistency.WriteRequest;
+import org.apache.rocketmq.mqtt.meta.raft.rpc.Constants;
 import org.apache.rocketmq.mqtt.meta.raft.snapshot.SnapshotOperation;
 import org.apache.rocketmq.mqtt.meta.raft.snapshot.impl.CounterSnapshotOperation;
